Organize files for fist counter letters exercioses
