+++ src/components/App/App.js
@@ -1,0 +1,10 @@
+import './App.css';
+
+export const App = () => {
+  return (
+    <div className="App">
+    </div>
+  );
+}
+
+export default App;
