+++ src/components/WordCounter/CharactherCounter.js
@@ -1,0 +1,4 @@
+const textState = atom({
+  key: 'textState',
+  default: '',
+})
