+++ src/index.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import { App }  from './components/App/App';
 import { RecoilRoot } from 'recoil';
 
-const recoilRoot = <RecoilRoot> <App /> </RecoilRoot>;
+const recoilRoot = <RecoilRoot><App /></RecoilRoot>;
 
 ReactDOM.render(recoilRoot ,document.getElementById('root'));
